+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## [Unreleased]
 
+## [v0.6.2]
+
 ### Added
 
 - Subcommand for downloading and installing Hyperspace via the CLI.
+- More FTL versions are now recognized correctly.
 
 ### Fixed
 
@@ -131,7 +134,8 @@
 - Added error context to .txt file decoding error.
 - Fixed XML Sandbox displaying the wrong file after the data archive's file list has been changed.
 
-[unreleased]: https://github.com/afishhh/ftlman/compare/v0.6.1...HEAD
+[unreleased]: https://github.com/afishhh/ftlman/compare/v0.6.2...HEAD
+[v0.6.2]: https://github.com/afishhh/ftlman/compare/v0.6.1...v0.6.2
 [v0.6.1]: https://github.com/afishhh/ftlman/compare/v0.6.0...v0.6.1
 [v0.6.0]: https://github.com/afishhh/ftlman/compare/v0.5.4...v0.6.0
 [v0.5.4]: https://github.com/afishhh/ftlman/compare/v0.5.3...v0.5.4